:recycle: move cli commands to cli files to remove index files
