+++ package.json
@@ -2,11 +2,10 @@
   "name": "@datadog/datadog-ci",
   "version": "0.17.8",
   "description": "Run datadog actions from the CI.",
-  "main": "dist/index.js",
   "repository": "https://github.com/DataDog/datadog-ci",
   "license": "Apache-2.0",
   "bin": {
-    "datadog-ci": "dist/index.js"
+    "datadog-ci": "dist/cli.js"
   },
   "engines": {
     "node": ">=10.24.1"
