+++ src/commands/dependencies/cli.ts
