--- src/commands/synthetics/index.ts
@@ -1,3 +1,0 @@
-import {RunTestCommand} from './cli'
-
-module.exports = [RunTestCommand]
