+++ src/commands/trace/cli.ts
